feat: add page pool for storing freed pages

+++ internal/storage/file-utils.go
@@ -5,10 +5,10 @@ import (
 	"syscall"
 )
 
-func increaseFileSize(file *os.File, size int) error {
-	return syscall.Fallocate(int(file.Fd()), 0, 0, int64(size))
+func increaseFileSize(file *os.File, size int64) error {
+	return syscall.Fallocate(int(file.Fd()), 0, 0, size)
 }
 
 func mapFileToMemory(file *os.File, offset int64, size int) (data []byte, err error) {
-	return syscall.Mmap(int(file.Fd()), offset, size, syscall.PROT_READ|syscall.PROT_WRITE, syscall.MAP_SHARED)
+	return syscall.Mmap(int(file.Fd()), int64(offset), int(size), syscall.PROT_READ|syscall.PROT_WRITE, syscall.MAP_SHARED)
 }
